# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Reto Tecnico Exchange Rate
  description: |
    Esta API permite realizar un tipo de cambio a un monto donde se deben utilizar el monto, moneda de origen, moneda de destino, monto con el tipo de campo y el tipo de cambio.
    
    Permite el registro, actualización y búsqueda del tipo de cambio.
    
    Por cada tipo de cambio realizado, contiene quien hizo la solicitud (auditoría funcional)

  contact:
    email: o.velezmoro@gmail.com.io
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/system-development/v1
tags:
  - name: ExchangeRate
    description: Tipo de Cambio
paths:
  /exchangeRate/:
    get:
      tags:
        - ExchangeRate
      summary: Búsqueda del tipo de cambio.
      operationId: exchangeRateGET
      parameters:
        - name: from
          in: query
          description: Moneda Origen
          required: true
          schema:
            type: string
            example: 'USD'
        - name: to
          in: query
          description: Moneda Destino
          required: true
          schema:
            type: string
            example: 'PEN'
      responses:
        200:
          description: Se obtuvó el recurso solicitado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateResponse"
      security:
        - bearerAuth: []
    post:
      tags:
        - ExchangeRate
      summary: Inserta un nuevo tipo de cambio
      operationId: createExchangeRatePOST
      requestBody:
        description: Petición del recurso
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRateRequest'
      responses:
        201:
          description: Se creo el recurso solicitado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateResponse"
    put:
      tags:
        - ExchangeRate
      summary: Actualiza un nuevo tipo de cambio
      operationId: updateExchangeRatePOST
      requestBody:
        description: Petición del recurso
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeRateRequest'
      responses:
        201:
          description: Se actualizo el recurso solicitado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRateResponse"
  /exchangeRate/convert:
    post:
      tags:
        - ExchangeRate
      summary: Convierte el tipo de cambio de un monto.
      operationId: convertExchangeRatePOST
      requestBody:
        description: Petición del recurso
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConvertExchangeRateRequest'
      responses:
        200:
          description: Se obtuvó el recurso solicitado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertExchangeRateResponse"
  /exchangeRate/history:
    get:
      tags:
        - ExchangeRate
      summary: Búsqueda el historial de un tipo de cambio.
      operationId: historyExchangeRateGET
      parameters:
        - name: date
          in: query
          description: Fecha del tipo de cambio
          required: true
          schema:
            type: string
            example: '2022-04-26'
        - name: from
          in: query
          description: Moneda Origen
          required: true
          schema:
            type: string
            example: 'USD'
      responses:
        200:
          description: Se obtuvó el recurso solicitado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryExchangeRateResponse"
      security:
        - bearerAuth: []
components:
  schemas:
    ExchangeRateRequest:
      type: object
      properties:
        baseCurrency:
          type: string
          description: Moneda Origen
          example: "USD"
        exchangeCurrency:
          type: string
          description: Código de Moneda.
          example: "PEN"
        rateAmount:
          type: number
          format: double
          description: Tipo de cambio de la moneda.
          example: 20.41006
    ExchangeRate:
      type: object
      properties:
        currencyCode:
          type: string
          description: Código de Moneda.
          example: "PEN"
        amount:
          type: number
          format: double
          description: Tipo de cambio de la moneda.
    ExchangeRateResponse:
      type: object
      properties:
        base:
          type: string
          description: Moneda Origen.
          example: "USD"
        result:
          description: Resultado de Monedas con su ratio de tipo de cambio.
          $ref: "#/components/schemas/ExchangeRate"
        updated:
          type: string
          description: Fecha en que se realizo la consulta.
          example: "2022-04-27 02:31:41"
    HistoryExchangeRateResponse:
      type: object
      properties:
        base:
          type: string
          description: Moneda Origen.
          example: "USD"
        result:
          type: array
          description: Resultado de Monedas con su ratio de tipo de cambio.
          items:
            $ref: "#/components/schemas/ExchangeRate"
        updated:
          type: string
          description: Fecha en que se realizo la consulta.
          example: "2022-04-27 02:31:41"
    ConvertExchangeRateRequest:
      type: object
      properties:
        from:
          type: string
          description: Moneda Origen.
          example: "USD"
        to:
          type: string
          description: Moneda Destino
          example: "PEN"
        amount:
          type: number
          format: double
          description: Monto a transformar
          example: 35.5
    ConvertExchangeRateResponse:
      type: object
      properties:
        base:
          type: string
          description: Moneda Origen.
          example: "USD"
        amount:
          type: number
          format: double
          description: Monto a transformar
          example: 35.5
        result:
          type: object
          description: Moneda Destino
          additionalProperties:
            type: number
            format: double
        updated:
          type: string
          description: Fecha en que se realizo la consulta.
          example: "2022-04-27 02:31:41"
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT